fix: lakera persistence issue
