+++ apps/nextjs/src/app/api/chat/webActionsPlugin.ts
@@ -1,6 +1,6 @@
 import type { AilaPlugin } from "@oakai/aila/src/core/plugins";
 import { AilaThreatDetectionError } from "@oakai/aila/src/features/threatDetection";
-import { handleThreatDetectionError } from "@oakai/aila/src/utils/moderation/moderationErrorHandling";
+import { handleThreatDetectionError } from "@oakai/aila/src/utils/threatDetection/threatDetectionHandling";
 import { inngest } from "@oakai/core/src/inngest";
 import { SafetyViolations as defaultSafetyViolations } from "@oakai/core/src/models/safetyViolations";
 import { UserBannedError } from "@oakai/core/src/models/userBannedError";
